requirements.txt
anthropic
supabase
python-dotenv
fastapi
uvicorn
pydantic
